Move struct yall_subsystem inside subsystem.c
